requirements.txt
absl-py==2.3.0
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
efficientnet==1.0.0
fastapi==0.115.14
flatbuffers==25.2.10
fonttools==4.58.4
gast==0.6.0
gitdb==4.0.12
gitpython==3.1.44
google-pasta==0.2.0
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
idna==3.10
image-classifiers==1.0.0
imageio==2.37.0
jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras-applications==1.0.8
kiwisolver==1.4.8
lazy-loader==0.4
libclang==18.1.1
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml-dtypes==0.5.1
namex==0.1.0
narwhals==1.44.0
networkx==3.5
numpy==2.1.3
opencv-python-headless==4.11.0.86
opt-einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pillow==11.3.0
pip==24.0
protobuf==5.29.5
pyarrow==20.0.0
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
scikit-image==0.25.2
scipy==1.16.0
segmentation-models==1.0.1
setuptools==80.9.0
simsimd==6.4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
streamlit==1.46.1
stringzilla==3.12.5
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tifffile==2025.6.11
toml==0.10.2
tornado==6.5.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
